Show server-side FPS on the web client.

+++ cpp_web_viz/common/common.hpp
@@ -7,7 +7,7 @@ namespace cpp_web_viz {
 extern double NANOSECONDS_PER_SECOND;
 extern double NANOSECONDS_PER_MILLISECOND;
 
-using Hz = double;
+using Hz = float;
 
 struct Position
 {
